+++ wp-content/themes/tests/js/html5.js
@@ -1,0 +1,7 @@
+
+			
+			function myFunction() {
+
+			    document.getElementById('myCheck').click();
+			}
+