+++ wp-content/themes/tests/js/navigation.js
@@ -1,0 +1,6 @@
+
+			
+			$('p').click(function(){
+			    alert('The paragraph was clicked.');
+			});
+